pyproject.toml
[tool.poetry]
name = "componentor"
version = "0.1.0"
description = ""
authors = ["ratushnyyvm <ratushnyyvm@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.1"
python-dotenv = "^1.0.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
isort = "^5.12.0"
ipython = "^8.13.2"
coverage = "^7.2.6"
factory-boy = "^3.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
